- Removed unused signature folder - Changed xades folder to original code from tobella - Fixed XML XAdES--EPES signing, now it works - Removed CR-Libre API calls for XML Signing - Added code to capture Haciendas API error messages to be displayed in invoice
