+++ cr_electronic_invoice/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
 from . import models
-from . import signature
 from . import extensions
 from . import wizard
