--- cr_electronic_invoice/signature/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import algorithms
-from . import constants
-from . import ns
-from . import template
-from .signature_context import SignatureContext
