--- cr_electronic_invoice/signature/ns.py
@@ -1,17 +1,0 @@
-# -*- coding: utf-8 -*-
-# © 2017 Creu Blanca
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-DSigNs = 'http://www.w3.org/2000/09/xmldsig#'
-DSigNs11 = 'http://www.w3.org/2009/xmldsig11#'
-DSignNsMore = 'http://www.w3.org/2001/04/xmldsig-more#'
-EncNs = 'http://www.w3.org/2001/04/xmlenc#'
-XPathNs = 'http://www.w3.org/TR/1999/REC-xpath-19991116'
-XPath2Ns = 'http://www.w3.org/2002/06/xmldsig-filter2'
-XPointerNs = 'http://www.w3.org/2001/04/xmldsig-more/xptr'
-Soap11Ns = 'http://schemas.xmlsoap.org/soap/envelope/'
-Soap12Ns = 'http://www.w3.org/2002/06/soap-envelope'
-NsExcC14N = 'http://www.w3.org/2001/10/xml-exc-c14n#'
-NsExcC14NWithComments = 'http://www.w3.org/2001/10/xml-exc-c14n#WithComments'
-
-NS_MAP = {'ds': DSigNs, 'ds11': DSigNs11}
