+++ cr_electronic_invoice/views/res_company_views.xml
@@ -6,20 +6,20 @@
 			<field name="model">res.company</field>
 			<field name="inherit_id" ref="base.view_company_form"/>
 			<field name="arch" type="xml">
-				
+
 				<!--<xpath expr="/form/*" position="before">
 					<header>
 						<field name="state" widget="statusbar"  nolabel="1" readonly="1" />
 					</header>
 				</xpath>-->
-				
+
 				<xpath expr="//sheet[1]" position="after">
 					<div class="oe_chatter">
 						<field name="message_follower_ids" widget="mail_followers" groups="base.group_user"/>
 						<field name="message_ids" widget="mail_thread"/>
 					</div>
 				</xpath>
-				
+
 
 				<field name="partner_id" position="after">
 					<field name="commercial_name"/>
@@ -32,7 +32,7 @@
 				<field name="state_id" position="attributes">
 					<attribute name="options">{"no_open": True,"no_create": 1, "no_create_edit": 1}</attribute>
 				</field>
-				
+
 				<field name="website" position="after">
 					<field name="county_id" domain="[('state_id', '=', state_id)]" options="{'no_open': 1,'no_create': 1, 'no_create_edit': 1}"/>
 					<field name="district_id" domain="[('county_id', '=', county_id)]" options="{'no_open': 1,'no_create': 1, 'no_create_edit': 1}"/>
@@ -65,7 +65,6 @@
 								<field name="frm_ws_identificador"/>
 								<field name="frm_ws_password" password="True"/>
 								<field name="frm_pin" password="True"/>
-								<field name="frm_ws_api"  widget="radio"/>
 								<!--<field name="pciaid"/>-->
 								<!--<field name="cta_id_electronic_invoice"/>-->
 							</group>
@@ -73,8 +72,7 @@
 								<field name="frm_callback_url"/>
 								<!--<field name="activated" readonly="True" />-->
 								<field name="frm_ws_ambiente"  widget="radio"/>
-                                <field name="frm_apicr_signaturecode"/>
-								<!--<field name="frm_tipo_id"/>-->
+            		<!--<field name="frm_tipo_id"/>-->
 							</group>
 						</group>
 					</page>
