+++ cr_electronic_invoice/xades/tobella_xades/constants.py
@@ -17,5 +17,5 @@ MAP_HASHLIB = {
     constants.TransformSha256: hashes.SHA256,
     constants.TransformSha384: hashes.SHA384,
     constants.TransformSha512: hashes.SHA512,
-    # constants.TransformRipemd160: hashes.RIPEMD160,
+#    constants.TransformRipemd160: hashes.RIPEMD160,
 }
